Updates tests for new database structure and minor error checking changes

+++ backend/lib/event_queries.py
@@ -8,7 +8,6 @@ from backend.lib.game_queries import get_game_id
 from backend.lib.helper_commands import check_admin_status, get_id_from_name, \
     get_id_from_title, get_game_name, AdminPermissionError, GameNotFoundError
 from mysql.connector.errors import IntegrityError
-
 from backend.lib.user_queries import UserNotFoundError
 
 
@@ -23,10 +22,9 @@ class EventQueries(commands.Cog):
     async def create_event(self, ctx, event_title, event_date, game_name, team_size):
         """
         Create new event
-        :param ctx: command parameter
         :param event_title: title of event
         :param event_date: date of event (formatted DD/MM/YYYY)
-        :param game_name: title of game to be played\
+        :param game_name: title of game to be played
         :param team_size: size of individual teams
         :return: new event table or error message
         """
@@ -41,7 +39,14 @@ class EventQueries(commands.Cog):
                 'team_size': team_size, # individual team size
             }
 
-            message = sql_create_event(data_insert, self.cursor, self.cnx)
+            sql_create_event(data_insert, self.cursor, self.cnx)
+
+            team_size = int(data_insert['team_size'])
+            teams = create_blank_teams(team_size)
+
+            embed = create_embed_message(data_insert['title'], data_insert['date'],
+                                         get_game_name(data_insert['game_id'], self.cursor),
+                                         teams)  # Created embeded message
         except DateFormatError:
             await ctx.send("Error: your date is invalid.  Please use MM/DD/YYYY format")
         except GameNotFoundError:
@@ -54,7 +59,7 @@ class EventQueries(commands.Cog):
             await ctx.send("Error: team size must be above 0")
         else:
             event_channel = self.bot.get_channel(self.event_channel_id)
-            msg = await event_channel.send(embed=message)
+            msg = await event_channel.send(embed=embed)
             sql_update_event_id(msg.id, event_title, self.cursor, self.cnx) # Set event_id in database
             await msg.add_reaction('☑')   # Add accept emoji to message
             await msg.add_reaction('🇽')    # Add decline emoji to message
@@ -65,23 +70,22 @@ class EventQueries(commands.Cog):
     async def delete_event(self, ctx, event_title):
         """
         Delete an event
-        :param ctx: command parameter
         :param event_title: title of event
         :return: new event table or error message
         """
 
         event_id = get_id_from_title(event_title, self.cursor)
-        if event_id == -1:  # Separate error checking for event (because we will need the id here)
-            await ctx.send("Error: the event you are attempting to delete does not exist")
-            return  # If no event found, return with error
+
         try:
             sql_delete_event(str(ctx.author), event_id, self.cursor, self.cnx)
         except AdminPermissionError:
             await ctx.send("Permission error: only admins may delete events")
+        except EventNotFoundError:
+            await ctx.send("Error: trying to delete an event that does not exist")
         else:
             sql_delete_all_registrations(event_id, self.cursor, self.cnx)   # Remove all registrations for event
-            event_channel = self.bot.get_channel(self.event_channel_id) # Get event channel
-            msg = await event_channel.fetch_message(event_id)  # Get event message
+            event_channel = self.bot.get_channel(self.event_channel_id)     # Get event channel
+            msg = await event_channel.fetch_message(event_id)       # Get event message
             await msg.delete()
 
             await ctx.send("Successfully deleted event " + event_title + "!")
@@ -90,91 +94,90 @@ class EventQueries(commands.Cog):
     async def get_events(self, ctx):
         """
         Get all events
-        :param ctx: command parameter
         :return: list of all scheduled events
         """
         await ctx.send(sql_get_events(self.cursor))
 
-    @commands.command()
-    async def create_registration(self, ctx, event_title):
-        """
-        Register for an event
-        :param ctx: command parameter
-        :param event_title: event's title
-        :return: new count of event registrations
-        """
-        event_id = get_id_from_title(event_title, self.cursor)
-        if event_id == -1:  # Separate error checking for event (because we will need the id here)
-            await ctx.send("Error: the event you are attempting to register for does not exist")
-            return  # If no event found, return with error
-        try:
-            sql_create_registration(str(event_id), str(ctx.author), self.cursor, self.cnx)
-        except UserNotFoundError:
-            await ctx.send("Error: player not found in database")
-        except ExistingRegistrationError:
-            await ctx.send("Error: player is already registered for this event")
-        else:
-            event_channel = self.bot.get_channel(self.event_channel_id)  # Get event channel
-            msg = await event_channel.fetch_message(event_id)  # Get event message
-
-            team_size = sql_get_team_size(event_id, self.cursor)  # Get size of teams from event
-            teams = get_teams_from_embed(msg.embeds[0], team_size)  # Get teams of event
-
-            # Place player in team, if error we return
-            if get_team_player_count(teams[0]) <= get_team_player_count(teams[1]):
-                teams[0] = add_player_to_team(teams[0], str(ctx.author))    # Attempt to add player to team
-                if teams[0] == -1:  # Teams are full
-                    await ctx.send("Error: can't add user to team, teams are full")
-                    return
-            else:
-                teams[1] = add_player_to_team(teams[1], ctx.author) # Add player to second team
-
-            # Update teams in event channel
-            embed = modify_embed_message_teams(msg.embeds[0], teams)
-            await msg.edit(embed=embed)
-
-            await ctx.send("Successfully registered " + str(ctx.author) + " for event " + event_title + "!")
-
-    @commands.command()
-    async def delete_registration(self, ctx, event_title):
-        """
-        Cancel a registration
-        :param ctx: command parameter
-        :param event_title: title of event to cancel registration for
-        :return: new count of event registrations
-        """
-
-        event_id = get_id_from_title(event_title, self.cursor)
-        if event_id == -1:  # Separate error checking for event (because we will need the id here)
-            await ctx.send("Error: the event you are attempting to remove your registration for does not exist")
-            return  # If no event found, return with error
-
-        try:
-            sql_delete_registration(event_id, str(ctx.author), self.cursor, self.cnx)
-        except UserNotFoundError:
-            await ctx.send("Error: player not found in database")
-        else:
-            event_channel = self.bot.get_channel(self.event_channel_id)  # Get event channel
-            msg = await event_channel.fetch_message(event_id)  # Get event message
-
-            team_size = sql_get_team_size(event_id, self.cursor)  # Get size of teams from event
-            teams = get_teams_from_embed(msg.embeds[0], team_size)  # Get teams of event
-
-            # Remove player from team, if error we return
-            tempTeam0 = remove_player_from_team(teams[0], str(ctx.author))
-            tempTeam1 = remove_player_from_team(teams[1], str(ctx.author))
-
-            if tempTeam0 == -1 and tempTeam1 == -1:
-                await ctx.send("Error: player is not in a team")
-            else:
-                if tempTeam0 != -1:
-                    teams[0] = tempTeam0
-                else:
-                    teams[1] = tempTeam1
-            embed = modify_embed_message_teams(msg.embeds[0], teams)
-            await msg.edit(embed=embed)
-
-            await ctx.send("Successfully removed " + str(ctx.author) + " from event " + event_title + "!")
+    # @commands.command()
+    # async def create_registration(self, ctx, event_title):
+    #     """
+    #     Register for an event
+    #     :param ctx: command parameter
+    #     :param event_title: event's title
+    #     :return: new count of event registrations
+    #     """
+    #     event_id = get_id_from_title(event_title, self.cursor)
+    #     if event_id == -1:  # Separate error checking for event (because we will need the id here)
+    #         await ctx.send("Error: the event you are attempting to register for does not exist")
+    #         return  # If no event found, return with error
+    #     try:
+    #         sql_create_registration(str(event_id), str(ctx.author), self.cursor, self.cnx)
+    #     except UserNotFoundError:
+    #         await ctx.send("Error: player not found in database")
+    #     except ExistingRegistrationError:
+    #         await ctx.send("Error: player is already registered for this event")
+    #     else:
+    #         event_channel = self.bot.get_channel(self.event_channel_id)  # Get event channel
+    #         msg = await event_channel.fetch_message(event_id)  # Get event message
+    #
+    #         team_size = sql_get_team_size(event_id, self.cursor)  # Get size of teams from event
+    #         teams = get_teams_from_embed(msg.embeds[0], team_size)  # Get teams of event
+    #
+    #         # Place player in team, if error we return
+    #         if get_team_player_count(teams[0]) <= get_team_player_count(teams[1]):
+    #             teams[0] = add_player_to_team(teams[0], str(ctx.author))    # Attempt to add player to team
+    #             if teams[0] == -1:  # Teams are full
+    #                 await ctx.send("Error: can't add user to team, teams are full")
+    #                 return
+    #         else:
+    #             teams[1] = add_player_to_team(teams[1], ctx.author) # Add player to second team
+    #
+    #         # Update teams in event channel
+    #         embed = modify_embed_message_teams(msg.embeds[0], teams)
+    #         await msg.edit(embed=embed)
+    #
+    #         await ctx.send("Successfully registered " + str(ctx.author) + " for event " + event_title + "!")
+    #
+    # @commands.command()
+    # async def delete_registration(self, ctx, event_title):
+    #     """
+    #     Cancel a registration
+    #     :param ctx: command parameter
+    #     :param event_title: title of event to cancel registration for
+    #     :return: new count of event registrations
+    #     """
+    #
+    #     event_id = get_id_from_title(event_title, self.cursor)
+    #     if event_id == -1:  # Separate error checking for event (because we will need the id here)
+    #         await ctx.send("Error: the event you are attempting to remove your registration for does not exist")
+    #         return  # If no event found, return with error
+    #
+    #     try:
+    #         sql_delete_registration(event_id, str(ctx.author), self.cursor, self.cnx)
+    #     except UserNotFoundError:
+    #         await ctx.send("Error: player not found in database")
+    #     else:
+    #         event_channel = self.bot.get_channel(self.event_channel_id)  # Get event channel
+    #         msg = await event_channel.fetch_message(event_id)  # Get event message
+    #
+    #         team_size = sql_get_team_size(event_id, self.cursor)  # Get size of teams from event
+    #         teams = get_teams_from_embed(msg.embeds[0], team_size)  # Get teams of event
+    #
+    #         # Remove player from team, if error we return
+    #         tempTeam0 = remove_player_from_team(teams[0], str(ctx.author))
+    #         tempTeam1 = remove_player_from_team(teams[1], str(ctx.author))
+    #
+    #         if tempTeam0 == -1 and tempTeam1 == -1:
+    #             await ctx.send("Error: player is not in a team")
+    #         else:
+    #             if tempTeam0 != -1:
+    #                 teams[0] = tempTeam0
+    #             else:
+    #                 teams[1] = tempTeam1
+    #         embed = modify_embed_message_teams(msg.embeds[0], teams)
+    #         await msg.edit(embed=embed)
+    #
+    #         await ctx.send("Successfully removed " + str(ctx.author) + " from event " + event_title + "!")
 
     @commands.command()
     async def query_event(self, ctx, event_name):
@@ -206,19 +209,15 @@ def sql_create_event(data_insert, cursor, cnx):
     if int(data_insert['team_size']) <= 0:
         raise TeamSizeError
 
-    team_size = int(data_insert['team_size'])
-    teams = create_blank_teams(team_size)
-
-    embed = create_embed_message(data_insert['title'], data_insert['date'],
-                                 get_game_name(data_insert['game_id'], cursor), teams) # Created embeded message
-
     cursor.execute('insert into event '
                    '(event_id, date, game_id, title, team_size) '
                    'values (%(event_id)s, %(date)s, %(game_id)s, '
-                   '%(title)s, %(team_size)s)', data_insert) # add new event
+                   '%(title)s, %(team_size)s)', data_insert)        # add new event
     cnx.commit()  # commit changes to database
 
-    return embed    # return embed message to send via events channel
+    cursor.execute('select * from event where title = %s', (data_insert['title'],))
+
+    return cursor.fetchall()    # return embed message to send via events channel
 
 
 def sql_update_event_id(event_id, title, cursor, cnx):
@@ -239,8 +238,14 @@ def sql_delete_event(auth_user, event_id, cursor, cnx):
     """
     check_admin_status(auth_user, True, cursor)  # see if the authorizing user is an admin
 
+    if len(sql_query_event(event_id, cursor)) == 0:
+        raise EventNotFoundError
+
     cursor.execute('delete from event where event_id = %s', (event_id, ))  # execute deletion query
     cnx.commit()  # commit changes to database
+
+    cursor.execute('select * from event where event_id = %s', (event_id,))
+    return cursor.fetchall()
 
 
 def sql_get_events(cursor):
@@ -331,8 +336,8 @@ def sql_get_team_size(event_id, cursor):
     return result[0][0]  # return event id
 
 
-def sql_query_event(title, cursor):
-    cursor.execute('select * from event where title = %s', (title,))
+def sql_query_event(event_id, cursor):
+    cursor.execute('select * from event where event_id = %s', (event_id,))
     return cursor.fetchall()
